* Moved classes to the api module * Added some docs
