+++ fabric/modules/bows/src/main/java/com/sigmundgranaas/forgero/bow/ForgeroBowInitializer.java
@@ -15,13 +15,13 @@ import com.sigmundgranaas.forgero.bow.item.BowGroupRegistrars;
 import com.sigmundgranaas.forgero.bow.item.DynamicBowItemRegistrationHandler;
 import com.sigmundgranaas.forgero.bow.predicate.BowPullPredicate;
 import com.sigmundgranaas.forgero.core.Forgero;
+import com.sigmundgranaas.forgero.core.api.identity.ModificationRuleRegistry;
 import com.sigmundgranaas.forgero.core.handler.HandlerBuilderRegistry;
 import com.sigmundgranaas.forgero.core.model.match.PredicateFactory;
 import com.sigmundgranaas.forgero.core.property.v2.attribute.attributes.AttackDamage;
 import com.sigmundgranaas.forgero.core.property.v2.attribute.attributes.Durability;
 import com.sigmundgranaas.forgero.core.property.v2.attribute.attributes.Weight;
 import com.sigmundgranaas.forgero.core.registry.RegistryFactory;
-import com.sigmundgranaas.forgero.core.state.identity.ModificationRuleRegistry;
 import com.sigmundgranaas.forgero.core.type.Type;
 import com.sigmundgranaas.forgero.fabric.api.entrypoint.ForgeroPreInitializationEntryPoint;
 import com.sigmundgranaas.forgero.minecraft.common.handler.use.StopHandler;
