+++ fabric/modules/bows/src/test/java/com/sigmundgranaas/forgero/bow/item/NamingRuleTest.java
@@ -8,16 +8,16 @@ import static com.sigmundgranaas.forgero.core.type.Type.SCHEMATIC;
 import java.util.List;
 import java.util.Optional;
 
+import com.sigmundgranaas.forgero.core.api.identity.DefaultRules;
+import com.sigmundgranaas.forgero.core.api.identity.ModificationRuleRegistry;
+import com.sigmundgranaas.forgero.core.api.identity.sorting.SortingRuleRegistry;
 import com.sigmundgranaas.forgero.core.property.Attribute;
 import com.sigmundgranaas.forgero.core.property.attribute.AttributeBuilder;
 import com.sigmundgranaas.forgero.core.property.v2.attribute.attributes.AttackDamage;
 import com.sigmundgranaas.forgero.core.state.Ingredient;
+import com.sigmundgranaas.forgero.core.state.NameCompositor;
 import com.sigmundgranaas.forgero.core.state.State;
 import com.sigmundgranaas.forgero.core.state.composite.Construct;
-import com.sigmundgranaas.forgero.core.state.identity.DefaultRules;
-import com.sigmundgranaas.forgero.core.state.identity.ModificationRuleRegistry;
-import com.sigmundgranaas.forgero.core.state.identity.NameCompositor;
-import com.sigmundgranaas.forgero.core.state.identity.sorting.SortingRuleRegistry;
 import com.sigmundgranaas.forgero.core.type.SimpleType;
 import com.sigmundgranaas.forgero.core.type.Type;
 import com.sigmundgranaas.forgero.core.util.SchematicMatcher;
