--- forgero-core/src/main/java/com/sigmundgranaas/forgero/core/state/identity/ModificationRule.java
@@ -1,11 +1,0 @@
-package com.sigmundgranaas.forgero.core.state.identity;
-
-import java.util.function.Function;
-
-import com.sigmundgranaas.forgero.core.state.State;
-
-public interface ModificationRule {
-	boolean applies(State state);
-
-	Function<String, String> transformation();
-}
